Avoid unnecessary unboxing
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@26451 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/org/starexec/app/RESTServices.java
@@ -4013,7 +4013,7 @@ public class RESTServices {
 	    // Update database with new permissions
 	    for(Integer permittedSpaceId : permittedSpaces){
 			if(permittedSpaceId != null){
-			    Permissions.set(userId,permittedSpaceId.intValue(),newPerm);
+			    Permissions.set(userId, permittedSpaceId, newPerm);
 			}
 	    }
